Make memory management 1000% better. It has just reached the quality "shit" :poop:
Restructured LibStorj so it automatically manages the instance of StorjEnv.
Changed the memory management for CStruct subtypes. They have now a list of allocated memory pointers which are freed in the deinit. This makes it easier to know what really has to be freed. Before that we tried to free memory even if it was allocated (and maybe even freed) by libstorj or any other third party. This is now fixed as just the pointers we add manually to the allocatedPointers list will be freed.